+++ extensions/music/src/add-to-library.tsx
@@ -1,11 +1,17 @@
 import { showHUD } from "@raycast/api";
-import { refreshCache, wait } from "./util/cache";
 import { addToLibrary } from "./util/scripts/player-controls";
-import { handleTaskEitherError } from "./util/utils";
+import { refreshCache, wait } from "./util/cache";
+import { pipe } from "fp-ts/lib/function";
+import * as TE from "fp-ts/TaskEither";
 
 export default async () => {
-  await handleTaskEitherError(addToLibrary)();
-  await showHUD("Added to Library");
-  await wait(5);
-  await refreshCache();
+  await pipe(
+    addToLibrary,
+    TE.map(async () => {
+      showHUD("Added to Library");
+      await wait(5);
+      await refreshCache();
+    }),
+    TE.mapLeft(() => showHUD("Failed to Add to Library"))
+  )();
 };