+++ extensions/music/src/util/utils.ts
@@ -1,16 +1,8 @@
-import { showToast, showHUD, Toast, getPreferenceValues, Clipboard, open } from "@raycast/api";
+import { showToast, showHUD, Toast, Clipboard, open, Color } from "@raycast/api";
 import * as E from "fp-ts/Either";
 import { pipe } from "fp-ts/lib/function";
 import * as TE from "fp-ts/lib/TaskEither";
 import { ScriptError } from "./models";
-
-export const AppleMusicColor = "#fb556d";
-export const MusicIcon = {
-  source: {
-    light: "../assets/music-light.svg",
-    dark: "../assets/music-dark.svg",
-  },
-};
 
 export const handleError = (error: Error) =>
   TE.tryCatch(() => showToast(Toast.Style.Failure, error.name, error.message), E.toError);
@@ -38,7 +30,7 @@ export const handleTaskEitherError = <T, E extends Error>(te: TE.TaskEither<E, T
             await open(
               "https://github.com/raycast/extensions/issues/new?assignees=&labels=extension%2C+bug&template=extension_bug_report.md&title=%5BMusic%5D"
             );
-            await showHUD(`Thanks for reporting this bug!`);
+            showHUD(`Thanks for reporting this bug!`);
           },
           shortcut: {
             key: "enter",
@@ -55,17 +47,6 @@ export const displayDuration = (duration: number) => {
   return `${hours > 1 ? `${hours} hours ` : hours == 1 ? "1 hour " : ""}${minutes} min`;
 };
 
-const preferences = getPreferenceValues();
-const { primaryPlaylistAction, secondaryPlaylistAction } = preferences;
-
-export const getPlaylistPrimaryAction = (): string => {
-  return primaryPlaylistAction;
-};
-
-export const getPlaylistSecondaryAction = (): string => {
-  return secondaryPlaylistAction;
-};
-
 export const getAttribute = (data: string, key: string) => {
   return data.includes(`${key}=`) ? data.split(`${key}=`)[1].split("$break")[0].replace("\n", "") : "";
 };
@@ -78,4 +59,52 @@ export const trimTitle = (title: string) => {
     title = title.padEnd(MAX_TITLE_LENGTH);
   }
   return title;
+};
+
+export const AppleMusicColor = "#fb556d";
+export const Icons = {
+  Music: {
+    source: {
+      light: "../assets/icons/music-light.svg",
+      dark: "../assets/icons/music-dark.svg",
+    },
+  },
+  Star: {
+    source: "../assets/icons/star.svg",
+    tintColor: Color.PrimaryText,
+  },
+  StarFilled: {
+    source: "../assets/icons/star-filled.svg",
+    tintColor: Color.PrimaryText,
+  },
+  Heart: {
+    source: "../assets/icons/heart.svg",
+    tintColor: Color.PrimaryText,
+  },
+  HeartFilled: {
+    source: "../assets/icons/heart-filled.svg",
+    tintColor: Color.Red,
+  },
+  Repeat: {
+    Off: {
+      source: "../assets/icons/repeat.png",
+      tintColor: Color.PrimaryText,
+    },
+    All: {
+      source: "../assets/icons/repeat.png",
+      tintColor: AppleMusicColor,
+    },
+    One: {
+      source: "../assets/icons/repeat-one.png",
+      tintColor: AppleMusicColor,
+    },
+  },
+  Album: {
+    source: "../assets/icons/album-icon.svg",
+    tintColor: Color.PrimaryText,
+  },
+  Playlist: {
+    source: "../assets/icons/playlist-icon.svg",
+    tintColor: Color.PrimaryText,
+  },
 };